Phase 2 step 6: Extract intraday (the _db call-time-resolution seam — riskiest helper move)
